LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/clearkey-mp4-reset-src-after-setmediakeys-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Media source error

Harness Error (FAIL), message = Media source error

PASS Reset src after setMediaKeys(). 
